requirements.txt
fastapi
uvicorn[standard]
tweepy
transformers
python-dotenv

